Refactor composite_template
