+++ examples/composite_template/example_application_window/composite_template.ui
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <interface>
-  <template class="ExApplicationWindow" parent="GtkApplicationWindow">
+  <template class="ExampleApplicationWindow" parent="GtkApplicationWindow">
     <property name="default_width">600</property>
     <property name="default_height">300</property>
     <property name="visible">True</property>
