+++ examples/composite_template/main.rs
@@ -1,79 +1,8 @@
-use glib::subclass::prelude::*;
-use gtk::{gio, glib};
-use gtk::{prelude::*, CompositeTemplate};
+mod example_application_window;
 
-mod imp {
-    use super::*;
-    use gtk::subclass::prelude::*;
+use gtk::prelude::*;
 
-    /// The private struct, which can hold widgets and other data.
-    #[derive(Debug, Default, CompositeTemplate)]
-    #[template(file = "composite_template.ui")]
-    pub struct ExApplicationWindow {
-        // The #[template_child] attribute tells the CompositeTemplate macro
-        // that a field is meant to be a child within the template.
-        #[template_child]
-        pub headerbar: TemplateChild<gtk::HeaderBar>,
-        #[template_child]
-        pub label: TemplateChild<gtk::Label>,
-        // You can specify the optional `id` argument if the id is not the same
-        // as the identifier
-        #[template_child(id = "subtitle_label")]
-        pub subtitle: TemplateChild<gtk::Label>,
-    }
-
-    #[glib::object_subclass]
-    impl ObjectSubclass for ExApplicationWindow {
-        const NAME: &'static str = "ExApplicationWindow";
-        type Type = super::ExApplicationWindow;
-        type ParentType = gtk::ApplicationWindow;
-
-        // Within class_init() you must set the template.
-        // The CompositeTemplate derive macro provides a convenience function
-        // bind_template() to set the template and bind all children at once.
-        fn class_init(klass: &mut Self::Class) {
-            Self::bind_template(klass);
-        }
-
-        // You must call `Widget`'s `init_template()` within `instance_init()`.
-        fn instance_init(obj: &glib::subclass::InitializingObject<Self>) {
-            obj.init_template();
-        }
-    }
-
-    impl ObjectImpl for ExApplicationWindow {
-        fn constructed(&self, obj: &Self::Type) {
-            obj.init_label();
-            self.parent_constructed(obj);
-        }
-    }
-
-    impl WidgetImpl for ExApplicationWindow {}
-    impl ContainerImpl for ExApplicationWindow {}
-    impl BinImpl for ExApplicationWindow {}
-    impl WindowImpl for ExApplicationWindow {}
-    impl ApplicationWindowImpl for ExApplicationWindow {}
-}
-
-glib::wrapper! {
-    pub struct ExApplicationWindow(ObjectSubclass<imp::ExApplicationWindow>)
-        @extends gtk::Widget, gtk::Window, gtk::ApplicationWindow, @implements gio::ActionMap, gio::ActionGroup;
-}
-
-impl ExApplicationWindow {
-    pub fn new<P: glib::IsA<gtk::Application>>(app: &P) -> Self {
-        glib::Object::new(&[("application", app)]).expect("Failed to create ExApplicationWindow")
-    }
-
-    pub fn init_label(&self) {
-        // To access fields such as template children, you must get
-        // the private struct.
-        let self_ = imp::ExApplicationWindow::from_instance(self);
-        self_
-            .subtitle
-            .set_text("This is an example window made using composite templates");
-    }
-}
+use example_application_window::ExampleApplicationWindow;
 
 fn main() {
     let application = gtk::Application::new(
@@ -83,7 +12,7 @@ fn main() {
     .expect("Failed to initialize application");
 
     application.connect_activate(|app| {
-        let win = ExApplicationWindow::new(app);
+        let win = ExampleApplicationWindow::new(app);
         win.show();
     });
